General refactoring to prepare for masking (#260)
* Added mask-attributions training loop

* Added configs and associated refactor for mask-attribs

* Fixed some args, docstrings, comments as per Dan comments on PR. Also changed Bool to Float for type of top-k mask.

* Further fixes for removing Bool and replacing with Float for topk mask type

* Fixed args as part of bug fix of type of topk mask tensor, Bool to Float.

* Changed the loss to nn.functional version and tidied up attrib_sparsity_loss calculation line

* Cleaned up multiline attrib_sparsity_loss calculation

* Added validation that mask attrib config is not none if attrib type is mask

* Added pydantic types to give tighter bounds on certain configs

* Deleted a rogue binary file

* Arg changes because CI failing

* Removed tests that are tested elsewhere

* fixed collect_subnetwork_attributions so that it can plot other types of attributions if they are used.

* Implemented measurement of correlations between attribs before and after update

* Revert "Merge main into mask_attrib"

This reverts commit e956c947676a578a08f488289579b5e96068a055, reversing
changes made to bebd058014e5827d8c446a590062e95c7d6ff1aa.

* Remove redundant topk masking (again)

* Add missing config arg to collect_subnetwork_attributions

* Add test_components for mask clarity

* Add post_relu_act_recon: bool to main config

* Formatting/docs updates

* Remove mask attribution calculation

* Remove check_attributions_update function

---------
