+++ .gitignore
@@ -3,7 +3,7 @@ neuronpedia_outputs/
 .env
 .vscode/settings.json
 
-wandb/
+**/wandb/
 .data/
 .checkpoints/
 # Byte-compiled / optimized / DLL files
