+++ spd/models/components.py
@@ -2,7 +2,7 @@ from typing import Any, Literal
 
 import einops
 import torch
-from jaxtyping import Bool, Float
+from jaxtyping import Float
 from torch import Tensor, nn
 
 from spd.hooks import HookPoint
@@ -83,7 +83,7 @@ class LinearComponent(nn.Module):
     def forward(
         self,
         x: Float[Tensor, "batch ... d_in"],
-        topk_mask: Bool[Tensor, "batch ... C"] | None = None,
+        topk_mask: Float[Tensor, "batch ... C"] | None = None,
     ) -> Float[Tensor, "batch ... d_out"]:
         """Forward pass through A and B matrices which make up the component for this layer.
 
@@ -98,7 +98,8 @@ class LinearComponent(nn.Module):
         # First multiply by A to get to intermediate dimension m
         inner_acts = einops.einsum(x, self.A, "batch ... d_in, ... C d_in m -> batch ... C m")
         if topk_mask is not None:
-            assert topk_mask.shape == inner_acts.shape[:-1]
+            # We could apply the mask after component_acts, but we do it here so our matrices become
+            # sparser and more efficient to compute with.
             inner_acts = einops.einsum(
                 inner_acts, topk_mask, "batch ... C m, batch ... C -> batch ... C m"
             )
@@ -107,6 +108,7 @@ class LinearComponent(nn.Module):
         component_acts = einops.einsum(
             inner_acts, self.B, "batch ... C m, ... C m d_out -> batch ... C d_out"
         )
+
         self.hook_component_acts(component_acts)
 
         # Sum over subnetwork dimension
