+++ spd/run_spd.py
@@ -2,193 +2,22 @@
 
 from collections.abc import Callable
 from pathlib import Path
-from typing import Any, ClassVar, Literal, Self
 
 import einops
 import matplotlib.pyplot as plt
 import torch
 import wandb
 from jaxtyping import Float
-from pydantic import (
-    BaseModel,
-    ConfigDict,
-    Field,
-    NonNegativeFloat,
-    PositiveFloat,
-    PositiveInt,
-    model_validator,
-)
 from torch import Tensor
 from torch.utils.data import DataLoader
 from tqdm import tqdm
 
+from spd.attributions import calculate_attributions
+from spd.configs import Config
 from spd.hooks import HookedRootModule
-from spd.log import logger
 from spd.models.base import SPDModel
 from spd.module_utils import collect_nested_module_attrs, get_nested_module_attr
-from spd.types import ModelPath, Probability
-from spd.utils import (
-    calc_recon_mse,
-    calc_topk_mask,
-    calculate_attributions,
-    get_lr_schedule_fn,
-    get_lr_with_warmup,
-)
-
-
-class TMSTaskConfig(BaseModel):
-    model_config = ConfigDict(extra="forbid", frozen=True)
-    task_name: Literal["tms"] = "tms"
-    feature_probability: Probability
-    train_bias: bool
-    bias_val: float
-    data_generation_type: Literal["exactly_one_active", "at_least_zero_active"] = (
-        "at_least_zero_active"
-    )
-    pretrained_model_path: ModelPath  # e.g. wandb:spd-tms/runs/si0zbfxf
-
-
-class ResidualMLPTaskConfig(BaseModel):
-    model_config = ConfigDict(extra="forbid", frozen=True)
-    task_name: Literal["residual_mlp"] = "residual_mlp"
-    feature_probability: Probability
-    init_scale: float = 1.0
-    data_generation_type: Literal[
-        "exactly_one_active", "exactly_two_active", "at_least_zero_active"
-    ] = "at_least_zero_active"
-    pretrained_model_path: ModelPath  # e.g. wandb:spd-resid-mlp/runs/j9kmavzi
-
-
-class Config(BaseModel):
-    model_config = ConfigDict(extra="forbid", frozen=True)
-    wandb_project: str | None = None
-    wandb_run_name: str | None = None
-    wandb_run_name_prefix: str = ""
-    seed: int = 0
-    topk: PositiveFloat | None = None
-    batch_topk: bool = True
-    exact_topk: bool = False
-    batch_size: PositiveInt
-    steps: PositiveInt
-    print_freq: PositiveInt
-    image_freq: PositiveInt | None = None
-    image_on_first_step: bool = True
-    slow_images: bool = False
-    save_freq: PositiveInt | None = None
-    lr: PositiveFloat
-    out_recon_coeff: NonNegativeFloat | None = None
-    act_recon_coeff: NonNegativeFloat | None = None
-    param_match_coeff: NonNegativeFloat | None = 1.0
-    topk_recon_coeff: NonNegativeFloat | None = None
-    schatten_coeff: NonNegativeFloat | None = None
-    schatten_pnorm: NonNegativeFloat | None = None
-    lp_sparsity_coeff: NonNegativeFloat | None = None
-    distil_from_target: bool = False
-    pnorm: PositiveFloat | None = None
-    C: PositiveInt
-    m: PositiveInt | None = None
-    lr_schedule: Literal["linear", "constant", "cosine", "exponential"] = "constant"
-    lr_exponential_halflife: PositiveFloat | None = None
-    lr_warmup_pct: Probability = 0.0
-    sparsity_loss_type: Literal["jacobian"] = "jacobian"
-    unit_norm_matrices: bool = False
-    attribution_type: Literal["gradient", "ablation", "activation"] = "gradient"
-    task_config: TMSTaskConfig | ResidualMLPTaskConfig = Field(..., discriminator="task_name")
-
-    DEPRECATED_CONFIG_KEYS: ClassVar[list[str]] = [
-        "topk_param_attrib_coeff",
-        "orthog_coeff",
-        "hardcode_topk_mask_step",
-        "pnorm_end",
-        "topk_l2_coeff",
-        "spd_type",
-        "sparsity_warmup_pct",
-    ]
-    RENAMED_CONFIG_KEYS: ClassVar[dict[str, str]] = {"topk_act_recon_coeff": "act_recon_coeff"}
-
-    @model_validator(mode="before")
-    def handle_deprecated_config_keys(cls, config_dict: dict[str, Any]) -> dict[str, Any]:
-        """Remove deprecated config keys and change names of any keys that have been renamed."""
-        # Move k from task_config to Config and rename it to C
-        if "task_config" in config_dict and "k" in config_dict["task_config"]:
-            logger.warning("task_config.k is deprecated, please use C in the main Config instead")
-            config_dict["C"] = config_dict["task_config"]["k"]
-            del config_dict["task_config"]["k"]
-
-        for key in list(config_dict.keys()):
-            val = config_dict[key]
-            if key in cls.DEPRECATED_CONFIG_KEYS:
-                logger.warning(f"{key} is deprecated, but has value: {val}. Removing from config.")
-                del config_dict[key]
-            elif key in cls.RENAMED_CONFIG_KEYS:
-                logger.info(f"Renaming {key} to {cls.RENAMED_CONFIG_KEYS[key]}")
-                config_dict[cls.RENAMED_CONFIG_KEYS[key]] = val
-                del config_dict[key]
-        return config_dict
-
-    @model_validator(mode="after")
-    def validate_model(self) -> Self:
-        # Check valid combinations of topk and batch_size
-        if self.topk is not None:
-            if self.batch_topk:
-                if not (self.batch_size * self.topk).is_integer():
-                    logger.warning(
-                        f"batch_size * topk={self.batch_size * self.topk} is not an integer, will "
-                        f"round down from {self.batch_size * self.topk} to "
-                        f"{int(self.batch_size * self.topk)} when calculating topk_mask"
-                    )
-            else:
-                if not self.topk.is_integer():
-                    raise ValueError("topk must be an integer when not using batch_topk")
-
-        # Warn if neither topk_recon_coeff nor lp_sparsity_coeff is set
-        if not self.topk_recon_coeff and not self.lp_sparsity_coeff:
-            logger.warning("Neither topk_recon_coeff nor lp_sparsity_coeff is set")
-
-        # If topk_recon_coeff is set, topk must be set
-        if self.topk_recon_coeff is not None:
-            assert self.topk is not None, "topk must be set if topk_recon_coeff is set"
-
-        # If lp_sparsity_coeff is set, pnorm must be set
-        if self.lp_sparsity_coeff is not None:
-            assert self.pnorm is not None, "pnorm must be set if lp_sparsity_coeff is set"
-
-        # Check that topk_recon_coeff is None if topk is None
-        if self.topk is None:
-            assert self.topk_recon_coeff is None, "topk_recon_coeff is not None but topk is"
-
-        # Give a warning if both out_recon_coeff and param_match_coeff are > 0
-        if (
-            self.param_match_coeff is not None
-            and self.param_match_coeff > 0
-            and self.out_recon_coeff is not None
-            and self.out_recon_coeff > 0
-        ):
-            logger.warning(
-                "Both param_match_coeff and out_recon_coeff are > 0. It's typical to only set one."
-            )
-
-        # If any of the coeffs are 0, raise a warning
-        msg = "is 0, you may wish to instead set it to null to avoid calculating the loss"
-        if self.topk_recon_coeff == 0:
-            logger.warning(f"topk_recon_coeff {msg}")
-        if self.lp_sparsity_coeff == 0:
-            logger.warning(f"lp_sparsity_coeff {msg}")
-        if self.param_match_coeff == 0:
-            logger.warning(f"param_match_coeff {msg}")
-
-        # Check that lr_exponential_halflife is not None if lr_schedule is "exponential"
-        if self.lr_schedule == "exponential":
-            assert (
-                self.lr_exponential_halflife is not None
-            ), "lr_exponential_halflife must be set if lr_schedule is exponential"
-
-        if self.schatten_coeff is not None:
-            assert (
-                self.schatten_pnorm is not None
-            ), "schatten_pnorm must be set if schatten_coeff is set"
-
-        return self
+from spd.utils import calc_recon_mse, calc_topk_mask, get_lr_schedule_fn, get_lr_with_warmup
 
 
 def get_common_run_name_suffix(config: Config) -> str:
@@ -459,17 +288,18 @@ def optimize(
             )
 
         post_weight_acts = {k: v for k, v in target_cache.items() if k.endswith("hook_post")}
+        pre_weight_acts = {k: v for k, v in target_cache.items() if k.endswith("hook_pre")}
         attributions = calculate_attributions(
             model=model,
+            config=config,
             batch=batch,
             out=out,
             target_out=target_out,
-            pre_weight_acts={k: v for k, v in target_cache.items() if k.endswith("hook_pre")},
+            pre_weight_acts=pre_weight_acts,
             post_weight_acts=post_weight_acts,
             component_acts={
                 k: v for k, v in spd_cache.items() if k.endswith("hook_component_acts")
             },
-            attribution_type=config.attribution_type,
         )
 
         lp_sparsity_loss_per_k = None
@@ -522,39 +352,31 @@ def optimize(
 
         act_recon_loss = None
         if config.act_recon_coeff is not None:
-            if isinstance(config.task_config, ResidualMLPTaskConfig):
-                # For now, we treat resid-mlp special in that we take the post-relu activations
-                # We ignore the mlp_out layers
-                assert layer_acts_topk is not None
-                post_relu_acts = {}
-                layer_acts_topk_after_relu = {}
-                for i in range(len(target_model.layers)):
-                    post_relu_acts[f"layers.{i}.mlp_in.hook_post"] = torch.nn.functional.relu(
-                        post_weight_acts[f"layers.{i}.mlp_in.hook_post"]
-                    )
-                    layer_acts_topk_after_relu[f"layers.{i}.mlp_in.hook_post"] = (
-                        torch.nn.functional.relu(layer_acts_topk[f"layers.{i}.mlp_in.hook_post"])
-                    )
-
-                act_recon_loss = calc_act_recon(
-                    target_post_weight_acts=post_relu_acts, layer_acts=layer_acts_topk_after_relu
-                )
-            else:
-                act_recon_layer_acts = (
-                    layer_acts_topk
-                    if layer_acts_topk is not None
-                    else {k: v for k, v in spd_cache.items() if k.endswith("hook_post")}
-                )
-                act_recon_loss = calc_act_recon(
-                    target_post_weight_acts=post_weight_acts,
-                    layer_acts=act_recon_layer_acts,
-                )
+            act_recon_layer_acts = (
+                layer_acts_topk
+                if layer_acts_topk is not None
+                else {k: v for k, v in spd_cache.items() if k.endswith("hook_post")}
+            )
+            target_post_weight_acts = post_weight_acts
+            if config.post_relu_act_recon:
+                relu = torch.nn.functional.relu
+                # Only do post-relu act recon for mlp_in layers and ignore the other layers
+                act_recon_layer_acts = {
+                    k: relu(v) for k, v in act_recon_layer_acts.items() if "mlp_in" in k
+                }
+                target_post_weight_acts = {
+                    k: relu(v) for k, v in target_post_weight_acts.items() if "mlp_in" in k
+                }
+            act_recon_loss = calc_act_recon(
+                target_post_weight_acts=target_post_weight_acts,
+                layer_acts=act_recon_layer_acts,
+            )
 
         if config.schatten_coeff is not None:
+            # Use the sparsity loss as the mask in the lp case, and topk_mask otherwise
             mask = topk_mask if topk_mask is not None else lp_sparsity_loss_per_k
             assert mask is not None
             schatten_pnorm = config.schatten_pnorm if config.schatten_pnorm is not None else 1.0
-            # Use the attributions as the mask in the lp case, and topk_mask otherwise
             schatten_loss = calc_schatten_loss(
                 As=collect_nested_module_attrs(model, attr_name="A", include_attr_name=False),
                 Bs=collect_nested_module_attrs(model, attr_name="B", include_attr_name=False),
@@ -641,7 +463,7 @@ def optimize(
 
         # Skip gradient step if we are at the last step (last step just for plotting and logging)
         if step != config.steps:
-            loss.backward()
+            loss.backward(retain_graph=True)
 
             if step % config.print_freq == 0 and config.wandb_project:
                 # Calculate gradient norm
