+++ tests/test_resid_mlp.py
@@ -4,15 +4,17 @@ import torch
 from jaxtyping import Float
 from torch import Tensor
 
+from spd.configs import Config
 from spd.experiments.resid_mlp.models import (
     ResidualMLPConfig,
     ResidualMLPModel,
     ResidualMLPSPDConfig,
     ResidualMLPSPDModel,
+    ResidualMLPTaskConfig,
 )
 from spd.experiments.resid_mlp.resid_mlp_dataset import ResidualMLPDataset
 from spd.module_utils import get_nested_module_attr
-from spd.run_spd import Config, ResidualMLPTaskConfig, optimize
+from spd.run_spd import optimize
 from spd.utils import DatasetGeneratedDataLoader, set_seed
 
 # Create a simple ResidualMLP config that we can use in multiple tests
