+++ tests/test_utils.py
@@ -5,12 +5,7 @@ import torch
 from jaxtyping import Float
 from torch import Tensor
 
-from spd.utils import (
-    SparseFeatureDataset,
-    calc_activation_attributions,
-    calc_topk_mask,
-    compute_feature_importances,
-)
+from spd.utils import SparseFeatureDataset, calc_topk_mask, compute_feature_importances
 
 
 def test_calc_topk_mask_without_batch_topk():
@@ -66,46 +61,6 @@ def test_calc_topk_mask_with_batch_topk_n_instances():
 
     result = calc_topk_mask(attribution_scores, topk, batch_topk=True)
     torch.testing.assert_close(result, expected_mask)
-
-
-def test_calc_activation_attributions_obvious():
-    component_acts = {"layer1": torch.tensor([[[1.0, 0.0], [0.0, 1.0]]])}
-    expected = torch.tensor([[1.0, 1.0]])
-
-    result = calc_activation_attributions(component_acts)
-    torch.testing.assert_close(result, expected)
-
-
-def test_calc_activation_attributions_different_d_out():
-    component_acts = {
-        "layer1": torch.tensor([[[1.0, 2.0], [3.0, 4.0]]]),
-        "layer2": torch.tensor([[[5.0, 6.0, 7.0], [8.0, 9.0, 10.0]]]),
-    }
-    expected = torch.tensor(
-        [[1.0**2 + 2**2 + 5**2 + 6**2 + 7**2, 3**2 + 4**2 + 8**2 + 9**2 + 10**2]]
-    )
-
-    result = calc_activation_attributions(component_acts)
-    torch.testing.assert_close(result, expected)
-
-
-def test_calc_activation_attributions_with_n_instances():
-    # Batch=1, n_instances=2, C=2, d_out=2
-    component_acts = {
-        "layer1": torch.tensor([[[[1.0, 2.0], [3.0, 4.0]], [[5.0, 6.0], [7.0, 8.0]]]]),
-        "layer2": torch.tensor([[[[9.0, 10.0], [11.0, 12.0]], [[13.0, 14.0], [15.0, 16.0]]]]),
-    }
-    expected = torch.tensor(
-        [
-            [
-                [1.0**2 + 2**2 + 9**2 + 10**2, 3**2 + 4**2 + 11**2 + 12**2],
-                [5**2 + 6**2 + 13**2 + 14**2, 7**2 + 8**2 + 15**2 + 16**2],
-            ]
-        ]
-    )
-
-    result = calc_activation_attributions(component_acts)
-    torch.testing.assert_close(result, expected)
 
 
 def test_dataset_at_least_zero_active():
